day14/requirement.txt
mysql-connector-python==9.4.0
python-dotenv==1.0.1